Create register/get method for global register lookups, so that finance API can move to local logged-in scope only. We want each API to have a specification where it is considered either global or local

+++ src/TAO/API/format.cpp
@@ -25,7 +25,7 @@ namespace TAO::API
     /* Outputs the correct balance in terms of a double based on decimals input. */
     double FormatBalance(const uint64_t nBalance, const uint8_t nDecimals)
     {
-        return (nBalance / math::pow(10, nDecimals));
+        return double(nBalance) / math::pow(10, nDecimals);
     }
